+++ rio_tiler/mosaic/backend.py
@@ -2,14 +2,18 @@
 
 import abc
 import logging
+import warnings
 from typing import Any
 
 import attr
+import numpy
 from morecantile import TileMatrixSet
 from pydantic import BaseModel, ConfigDict, Field
 from rasterio.crs import CRS
+from rasterio.errors import NotGeoreferencedWarning
 from rasterio.features import bounds as featureBounds
-from rasterio.features import geometry_mask
+from rasterio.features import rasterize
+from rasterio.warp import transform_geom
 
 from rio_tiler.constants import WEB_MERCATOR_TMS, WGS84_CRS
 from rio_tiler.errors import NoAssetFoundError, PointOutsideBounds
@@ -220,20 +224,47 @@ class BaseBackend(BaseReader):
         self,
         shape: dict,
         shape_crs: CRS = WGS84_CRS,
+        dst_crs: CRS | None = None,
         search_options: dict | None = None,
         **kwargs: Any,
     ) -> tuple[ImageData, list[str]]:
         """Create an Image from multiple assets for a GeoJSON feature."""
         shape = _validate_shape_input(shape)
+
+        if not dst_crs:
+            dst_crs = shape_crs
+
         bbox = featureBounds(shape)
 
         img, asset_used = self.part(
             bbox,
             bounds_crs=shape_crs,
+            dst_crs=dst_crs,
             search_options=search_options,
             **kwargs,
         )
-        img.array.mask = geometry_mask([shape], (img.height, img.width), img.transform)
+
+        if dst_crs != shape_crs:
+            shape = transform_geom(shape_crs, dst_crs, shape)
+
+        with warnings.catch_warnings():
+            warnings.filterwarnings(
+                "ignore",
+                category=NotGeoreferencedWarning,
+                module="rasterio",
+            )
+            cutline_mask = rasterize(
+                [shape],
+                out_shape=(img.height, img.width),
+                transform=img.transform,
+                all_touched=True,  # Mandatory for matching masks at different resolutions
+                default_value=0,
+                fill=1,
+                dtype="uint8",
+            ).astype("bool")
+
+        img.array.mask = numpy.where(~cutline_mask, img.array.mask, True)
+
         return img, asset_used
 
     ############################################################################